Feature2

+++ src/main/java/com/nurmanhilman/eventbrite/controller/PromotionController.java
@@ -3,23 +3,34 @@ package com.nurmanhilman.eventbrite.controller;
 
 import com.nurmanhilman.eventbrite.entities.EventEntity;
 import com.nurmanhilman.eventbrite.entities.PromotionEntity;
+import com.nurmanhilman.eventbrite.entities.UserEntity;
+import com.nurmanhilman.eventbrite.exception.CustomResponseStatusException;
+import com.nurmanhilman.eventbrite.repositories.EventRepository;
 import com.nurmanhilman.eventbrite.service.PromotionService;
+import com.nurmanhilman.eventbrite.service.UserService;
 import org.apache.coyote.Response;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.time.Instant;
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/api/v1/promotions")
 public class PromotionController {
 
     private final PromotionService promotionService;
+    private final UserService userService;
+    private final EventRepository eventRepository;
 
     @Autowired
-    public PromotionController(PromotionService promotionService) {
+    public PromotionController(PromotionService promotionService, UserService userService, EventRepository eventRepository) {
+        this.userService = userService;
         this.promotionService = promotionService;
+        this.eventRepository = eventRepository;
     }
 
     @GetMapping
@@ -36,7 +47,17 @@ public class PromotionController {
     }
 
     @PostMapping
-    public PromotionEntity createPromotion(@RequestBody PromotionEntity promotion) {
+    public PromotionEntity createPromotion(@RequestHeader("Authorization") String authorizationHeader,
+                                           @RequestBody PromotionEntity promotion) {
+        UserEntity userEntity = userService.getUserFromJwt(authorizationHeader);
+        System.out.println("Event id = "+ promotion.getEventId());
+        Optional<EventEntity> eventEntity = eventRepository.findById(promotion.getEventId());
+        if (eventEntity.isEmpty())
+            throw new CustomResponseStatusException(HttpStatus.NOT_FOUND, "Event id " + promotion.getEventId() +" is not found");
+        if (eventEntity.get().getUserId() != userEntity.getUserId())
+            throw new CustomResponseStatusException(HttpStatus.FORBIDDEN, "This user id is not owner of the event id "+promotion.getEventId());
+        promotion.setCreatedAt(Instant.now());
+        promotion.setUpdatedAt(Instant.now());
         return promotionService.createPromotion(promotion);
     }
 
@@ -51,7 +72,11 @@ public class PromotionController {
     }
 
     @DeleteMapping("/{promoId}")
-    public ResponseEntity<Void> deletePromotion(@PathVariable Long promoId) {
+    public ResponseEntity<Void> deletePromotion(@RequestHeader("Authorization") String authorizationHeader,
+                                                @PathVariable Long promoId) {
+        UserEntity userEntity = userService.getUserFromJwt(authorizationHeader);
+        if (!promotionService.isUserOwnerOfPromotion(promoId, userEntity.getUserId()))
+            throw new CustomResponseStatusException(HttpStatus.FORBIDDEN, "This user id is not owner of the promo id "+promoId);
         promotionService.deletePromotion(promoId);
         return ResponseEntity.noContent().build();
     }
